ci: add usage guard to build workflow
A usage guard job has been added to the GitHub Actions build workflow. This job utilizes the nekofar/action-usage-guard action to monitor resource consumption during the build process. It is set to halt the build if the usage exceeds 70% of the allotted resources. The addition of this guard enhances the reliability and efficiency of the build workflow.