ci: replace GitHub token command with nekofar action for draft release
This commit replaces the command format using GITHUB_TOKEN in the GitHub Actions build workflow with the nekofar/create-github-release@v1.0.8 action for creating a release draft. It is intended to streamline the release drafting process using a predefined GitHub Action. The new setup also allows automatic marking of the draft as a pre-release depending on the version tag, enhancing the versatility of the action.