build: upgrade Gradle version to 8.2.1
Updated the Gradle version from 8.1.1 to 8.2.1 in both gradle.properties and gradle-wrapper.properties files. This was necessary to take advantage of the improvements and bug-fixes introduced in the newer version of Gradle. Also, a validateDistributionUrl property was added to enhance the security of the Gradle distribution.

fixup! build: upgrade Gradle version to 8.2.1